move utils; add atmallen user
