+++ elk_generalization/datasets/popqa_weak_lm_dataset.py
@@ -3,8 +3,8 @@ from collections import defaultdict
 
 from datasets import Dataset, load_dataset
 
-from ..utils import assert_type
-from .weak_lm_dataset import QADataset
+from utils import assert_type
+from weak_lm_dataset import QADataset
 
 # from https://github.com/AlexTMallen/adaptive-retrieval
 WEAK_LM_TEMPLATE = "Q: {question} A:"
