+++ elk_generalization/datasets/sentiment_dataset.py
@@ -4,8 +4,8 @@ from collections import defaultdict
 
 from datasets import Dataset, concatenate_datasets, load_dataset
 
-from ..utils import transpose_dict
-from .quirky_dataset import QuirkyDataset
+from utils import transpose_dict
+from quirky_dataset import QuirkyDataset
 
 ZERO_SHOT_TEMPLATE = (
     'Title: {title}\n"""{review}"""\nQ: Is this review Positive or Negative?\nA:'
