+++ elk_generalization/datasets/weak_lm_dataset.py
@@ -3,8 +3,8 @@ from collections import defaultdict
 
 import numpy as np
 
-from ..utils import transpose_dict
-from .quirky_dataset import QuirkyDataset
+from utils import transpose_dict
+from quirky_dataset import QuirkyDataset
 
 
 class WeakLMDataset(QuirkyDataset):
@@ -20,7 +20,7 @@ class WeakLMDataset(QuirkyDataset):
     def __init__(self, weak_model_name: str = "EleutherAI/pythia-410m", **kwargs):
         super().__init__(**kwargs)
         weak_model_last = weak_model_name.split("/")[-1]  # TODO: better way to do this?
-        self.dataset_name += f"_{weak_model_last}"
+        self.name += f"_{weak_model_last}"
         self.weak_model_name = weak_model_name
 
     def _generate_base_dataset(
